+++ src/sdk/product/useProduct.ts
@@ -20,7 +20,7 @@ export const useProduct = <T extends BrowserProductQueryQuery>(
   productID: string,
   fallbackData?: T
 ) => {
-  const { channel } = useSession()
+  const { channel, locale } = useSession()
   const variables = useMemo(() => {
     if (!channel) {
       throw new Error(`useProduct: 'channel' from session is an empty string.`)
@@ -30,9 +30,10 @@ export const useProduct = <T extends BrowserProductQueryQuery>(
       locator: [
         { key: 'id', value: productID },
         { key: 'channel', value: channel },
+        { key: 'locale', value: locale },
       ],
     }
-  }, [channel, productID])
+  }, [channel, locale, productID])
 
   return useQuery<
     BrowserProductQueryQuery & T,