+++ store.config.js
@@ -11,6 +11,7 @@ module.exports = {
 
   // Default channel
   channel: '{"salesChannel":"1"}',
+  locale: 'en-US',
 
   // Production URLs
   storeUrl: 'https://vtexfaststore.com',